Fix lightweight camera test imports
